Improve type safety in VectorF_FromVector function
Explicitly cast vector components to float before using
floorf to enhance clarity and prevent implicit conversions.

+++ StrangeAwakening/Src/VectorF.c
@@ -67,5 +67,5 @@ VectorF VectorF_Normalize(VectorF vector)
 
 VectorF VectorF_FromVector(Vector vector)
 {
-    return VectorF_New((int)floorf(vector.x), (int)floorf(vector.y));
+    return VectorF_New(floorf((float)vector.x), floorf((float)vector.y));
 }
